[Xamarin.Android.Build.Tasks] Add `AndroidAarLibrary` Build Action
Our components team currently have to go through a ton
of hoops to include .aar files in the build system.
Because we are not allowed to ship some of the google .aar
files with the Nuget packasges they need to be downloaded
on demand. To work within the current build system these
.aar files then need to be unpacked, restructured,
compressed and included in the Nuget assembly as an
EmbeddedResource. This requires updating the assembly
after it has been downloaded from nuget.

With the introduction of the `AndroidAarLibrary` build
action we can remove the rebuilding of the Nuget assembly
completely. The new system will still download the required
.aar files on demand. But instead of repacking them into
the .dll, the Nuget .targets file will simply add these
.aar files to the `AndroidAarLibrary` ItemGroup. These
Items will then be picked up by the `ResolveLibraryProjectImports`
task. They will then be extracted into the `$(IntermediateOutputDir)`.
From there the resources, assets and .jar files in the .aar will
then be included in the build system of the application.