[Xamarin.Android.Build.Tasks] LogicalName & __AndroidLibraryProjects__ (#1239)
[Xamarin.Android.Build.Tasks] LogicalName & __AndroidLibraryProjects__

Fixes: https://github.com/xamarin/xamarin-android/issues/1235

There seems to be a [bug in msbuild][1] where if you create an item
inside a task and the task is skipped, the metadata is not created
for the item.

We get into a situation where `__AndroidLibraryProjects__.zip`
does not contain the correct `%(EmbededResource.LogicalName)` value.
As a result the extraction process cannot find the resource, and
builds fail.

Fix this by splitting out the code that created the zip file into a
new `_CreateManagedLibraryResourceArchive` target. This allows us to
skip this new target if needed. However the
`_AddLibraryProjectsEmbeddedResourceToProject` target which creates
the item can now run for every build. This will ensure that the
metadata is always created.

Add a unit test.

[1]: https://social.msdn.microsoft.com/Forums/netframework/en-US/259271d8-a3fc-4a9b-9e3a-fecad8f6f63f/potential-bug-with-createitem-task-and-additionalmetadata?forum=msbuild