Fix build with msbuild on Linux (#1270)
XA's Makefile exports the OS environment variable set to output of the `uname
-s` command which returns the operating system's name. On macOS it will return
`Darwin`, on Linux `Linux` etc. MSBuild (the build system, not the command)
defines a standard property `$(OS)` which names the operating system kind -
`Windows_NT` on Windows and `Unix` on *nix systems.

Our project and target files (also a large set of them from e.g. Mono) use the
property to check whether they are running on Unix or Windows and they rely on
the property's value to be set correctly to make the right decisions.
MSBuild (the system) supports setting of properties from the command line but it
also takes their values from the process environment. However, it appears
there's a major difference between Mono's xbuild (which is used by default by
XA) and msbuild in that xbuild "protects" the `OS` property by NOT using the
OS environment property to set its value, while msbuild sets the property from
the environment variable. This results in the property to have the value of
`Darwin` on macOS and `Linux` on Linux and it happens to work (more or less)
correctly on macOS because we do check for `'$(OS)' == 'Darwin' or '$(OS)' ==
'Unix')` in a few places while we do NOT check for `'$(OS)' == 'Linux')`
anywhere. All this together breaks XA build on Linux in a few places.

This patch makes the msbuild build (`make MSBUILD=msbuild`) to work on Linux by
renaming the Make `$(OS)` variable to `$(CURRENT_OS)` and thus "protecting" the
standard value of the `$(OS)` property during the build.