Normalize paths passed to api-merge
Depending on whether api-merge is invoked by xbuild or msbuild, the paths passed
to the app can be relative, full or mixed. However, the code assumes that
they're all equal and compares them to know when to stop processing the passed
sources.

Since the assumption is wrong and it breaks our builds, this patch makes sure
that all paths are normalized before they are compared.