[create-vsix] Remove unnecessary cruft
Fixes: https://github.com/xamarin/xamarin-android/issues/1267

The commercial Windows `Xamarin.Android.Sdk.*.vsix` file contains lots
of files that the macOS `xamarin.android*.pkg` file does not.

Update `build-tools/create-vsix` to remove some of the additional
files which are not necessary, including:

  * `*.dylib`: macOS native libraries, not needed on Windows.
  * `$MSBuild/Xamarin/Android/**/*.pdb`: Debug symbols.
  * `$MSBuild/Xamarin/Android/lib/**/*.d.so`: Native libraries which
    have not been `strip`d.
  * `$ReferenceAssemblies/Microsoft/Framework/MonoAndroid/**/libZipSharp.*
    `libZipSharp*` is only needed in `$MSBuild`.
  * Distribute `Mono.Android.Export.*`, `OpenTK.*`, and `OpenTK-1.0.*`
    for only *one* API level, not *every* API level.
    `RedistList/FrameworkList.xml` ensures that previous framework
    versions will be included in assembly resolution via
    `/FileList/@IncludeFramework`, so there is no need to ship 10+
    identical copies of these assemblies.

A remaining (significant!) contribution to the size difference between
Windows and macOS is *documentation*: `Mono.Android.xml` is included
for each `$(TargetFrameworkVersion)`, and is ~57MB *per copy*.
This will *not* be addressed in xamarin-android, as documentation is
handled elsewhere (and I'm not entirely sure how to handle it).