[Xamarin.Android.Build.Tasks] Add GetAndroidDependencies Target
Fixes #1269

This commit adds the `GetAndroidDependencies` target to the
`Xamarin.Android.Common.targets`. Its purpose is to examine
the various settings in the project and report which android
sdk build-tools, platform-tools etc are required.

`GetAndroidDependencies` will output an @(AndroidDependency)
with Version metadata Valid component names are platform,
build-tool, platform-tool and tool.