[build] fix api-xml-adjuster.targets on Windows
Since 7d705bf, the Windows builds on VSTS have seemed to be failing. I
tested this locally, and noticed the use of command line arguments such
as:

    -parameter-names='%(SomeVariable)'

Unfortunately, this isn't working on Windows due to the single quote. It
is more appropriately expressed as:

    -parameter-names=&quot;%(SomeVariable)&quot;

This isn't very pretty, but it should work on all platforms.

I also fixed all the tabs I saw in this file--in favor of spaces, and
fixed other XML code conventions.

I will now return to my regularly scheduled baby duty.