[Xamarin.Android.Build.Tasks] Only report `ndk-bundle` if required.
We should only list `ndk-bundle` in the AndroidDependencies if
the user is using Aot or MkBundle. Otherwise we will end up
downloading a ton of stuff we don't need.

This commit fixes the `CalculateProjectDependencies` to report
`ndk-bundle` if needed. It also adds a unit test to make sure
it is not included when it shouldn't be.