[BCL-Tests] Remove '-s' from _GrantPermissions (#1299)
The [semantics of the `$(AdbTarget)` property][adb-target] are the
same as the [`adb` Target Device option][adb], which permits any of:

[adb-target]: Documentation/build_process.md
[adb]: https://developer.android.com/studio/command-line/adb.html#issuingcommands

  * `-d`: Only attached *device*.
  * `-e`: Only attached *emulator*
  * `-s SERIAL_NUMBER`: A specifically named target; needed if there
    is more than one attached device or emulator.

The problem with commit c4e81655 is that it overrode the
`$(InstallDependsOnTargets)` property to call the `_GrantPermissions`
target, which doesn't properly use `$(AdbTarget)`; it instead
*requires* using the `-s` option, which is inconsistent:

	adb -s $(AdbTarget) shell pm grant ...

Remove `-s` from the `adb` invocation so that `$(AdbTarget)` can
contain e.g. `-d` or `-e`, as is intended & documented.