[Mono.Android] remove obsoleted annotation-related metadata.
We used to bind Java annotation interfaces in different way. After we
changed them to the current style, we couldn't remove them immediately
to keep API compatibility.

Now we don't have to, it happened many years ago, and it's time to reduce
technical debt in massive amount of lines in metadata fixup.