[Xamarin.Android.Build.Tasks] GetAndroidDependencies should call _SetLatestTargetFrameworkVersion
This is a bug in the `<GetAndroidDependencies/>` target. It does
not call `MonoAndroidHelper.RefreshSupportedVersions`. As a result
if called directly it will generate a Null Reference exception because
`MonoAndroidHelper.SupportedVersions` will be null.

So we should be dependent on `<_SetLatestTargetFrameworkVersion/>`.
This will initialize the `MonoAndroidHelper.SupportedVersions`. But
it will also resolve the correct `TargetFrameworkVersion` if the
user is using `AndroidUseLatestPlatformSdk`.