[Xamarin.Android.Build.Tasks] GetAndroidDependencies calls _SetLatestTargetFrameworkVersion (#1305)
This is a bug in the `GetAndroidDependencies` target: it does not
cause `MonoAndroidHelper.RefreshSupportedVersions()` to be invoked.
As a result if called directly it will generate a
`NullReferenceException` as `MonoAndroidHelper.SupportedVersions`
will be `null`.

The `GetAndroidDependencies` should thus depend upon the
`_SetLatestTargetFrameworkVersion` target. This will initialize
`MonoAndroidHelper.SupportedVersions`, and will also resolve the
correct `$(TargetFrameworkVersion)` if the user has
`$(AndroidUseLatestPlatformSdk)`=True.