[Xamarin.Android.Build.Tasks] Stop designtime designer being deleted.
Fixes #1286

We have a number of problems with our DesignTime build system. The
main one which this PR addresses is the designer file is deleted
by `IncrementalClean` AND `CoreClean`. This them completely messes
up the DesignTime system since it can no longer find the file.

So what we should be doing is making sure we don't tell the build
system about the designer file ;). We do this by not writing the
path to the `$(CleanFile)`.