[Xamarin.Android.Build.Tasks] Stop designtime designer being deleted. (#1307)
Fixes: https://github.com/xamarin/xamarin-android/issues/1286

We have a number of problems with our DesignTime build system. The
main one which this PR addresses is the designer file is deleted
by `IncrementalClean` AND `CoreClean`. This them completely messes
up the DesignTime system since it can no longer find the file.

So what we should be doing is making sure we don't tell the build
system about the designer file ;-). We do this by not writing the
path to the `$(CleanFile)`.

~~ Some background on the [DesignTime build system][2] ~~

The primary entry point is the `CoreCompile` target, with occasional
calls to the `RefreshReferences` target. In theory
[`$(DesignTimeBuild)`][1] should be set to `true`; however, this is
not always the case which is why we have the `_SetupForDesignTimeBuild`
target which hooks into the `CoreCompile` target. Additionally,
`$(BuildingProject)` should be False. ([Additional information][2]).
However this document is not seem to be 100% accurate since the IDE's
(namely Visual Studio) do not always set the required properties. One
of the key requirements for design time builds seems to be that they
are independent of the main builds. The files it uses should not be
removed since the `CoreCompile` target does not seem to be called
consistently. For example, if the designer file is removed and the
`RefreshReferences` target is called, the design time build may well
fail because the targets required to re-generate the designer file
are not called.

Now for the problem: the `IncrementalClean` target runs between builds.
It seems to want to remove files from the previous build. I suspect the
purpose is to remove files that are no longer needed. (For example an
assembly which is no longer referenced.) The problem is that we have
a file (the design time `Resource.designer.cs`) which is not built as
part of the normal build. But this file was still being written to
`$(CleanFile)` which is the driver for the `IncrementalClean` process.
As a result, the file gets removed because its not in the list of
current `@(FileWrites)`. Not writing the name of this file to
`$(CleanFile)` seems to hide it from `IncrementalClean`.

There was also another problem with the `CoreClean` target. This was
also removing the designer files. Again this was down to the fact
that the file was listed in `$(CleanFile)`. However our older
workarounds did not work since they relied on `IncrementalClean`
running. For a `Clean` target invocation it does not. Again the
solution seems to be Don't Write the file to `$(CleanFile)`, which is
counterintuitive since we are told to write all files we generate to
that file.

[1]:  https://github.com/dotnet/project-system/blob/master/docs/design-time-builds.md#determining-whether-a-target-is-run-in-a-design-time-build
[2]: https://github.com/dotnet/project-system/blob/master/docs/design-time-builds.md