[msbuild] Fix <Import>-ed filename, for case sensitivity.
There is a trivial difference between `msbuild` and `xbuild` - `msbuild`
seems to respect (or ignorant of) the fact that filenames that are different
in case are actually different on case-sensitive filesystem.

Thus with `msbuild`, our binding projects don't build due to "missing"
Microsoft.CSharp.Targets (there are only Microsoft.CSharp.targets even in
`xbuild` support files in mono).

This trivial fix should make it work.