[Xamarin.Android.Build.Tasks] `$(AndroidUseLatestPlatformSdk)` uses STABLE levels (#1228)
Fixes: https://github.com/xamarin/xamarin-android/issues/1221

Currently when a user sets `$(AndroidUseLatestPlatformSdk)` it will
automatically use the maximum known API level. This is not always a
stable ApiLevel, e.g. when a preview API level has been bound, and
the binding has been installed.

Users will generally want to use the maximum *stable* API level when
`$(AndroidUseLatestPlatformSdk)` is True.

Alters things so that we use the maximum Stable API level when
`$(AndroidUseLatestPlatformSdk)` is True.

However, allow users to use unstable API levels if desired. This can
be achieved by setting `$(TargetFrameworkVersion)` to the desired
unstable framework version. (For example, when API-O was in preview,
this would be `v7.99`. When API-P is presumably announced, it will be
bound as `v8.99`.)

This will require a *manual* change to the `.csproj`, as we do not
anticipate the IDEs showing unstable API levels (though this could
plausibly be done with sufficient warnings...).