[Xamarin.Android.Build.Tasks] Add Default Metadata for `@(AndroidResource)` Items. (#1345)
Fixes: https://github.com/xamarin/xamarin-android/issues/1287

One of the problems we have encountered is that the logic behind how
and when the Design-Time Build (DTB) is run is unclear. Specifically
what events/changes trigger a DTB. After lengthy conversations with
the Visual Studio team it seems that their preferred solution is to
use [`%(Generator)` metadata][0].

The idea behind this is that an Item can include

	<AndroidResource Include="Resources\layout\Main.axml">
	  <Generator>MSBuild:Target</Generator>
	</AndroidResource>

metadata which states that the MSBuild `Target` should be run when
the `%(Identity)` file is modified. When the `%(Identity)` file is
modified within the IDE, the specified MSBuild target will be
executed as part of a new Design-Time Build.

Normally you define the `%(Generator)` metadata on the Item in the
`.csproj`. However, We have *lots* of developers with existing
projects which DO NOT include this metadata. How do we fix up
existing projects without a) requiring manual changes or b) having to
write a tool to upgrade the projects?

This is where [`<ItemDefinitionGroup/>`][1] comes in :). It allows
us to define the `default` metadata we want to include on Items.
If we have:

	<ItemDefinitionGroup>
	  <Foo>
           <A>bar</A>
         </Foo>
	</ItemDefinitionGroup>

	<Foo Include="1">
	  <A>some</A>
	</Foo>
	<Foo Include="2" />

What will happen is `<A>bar</A>` will be automatically added to
the `<Foo Include="2" />`. It will be skipped on the other one
because it already has a `<A/>` element defined. This is a really
nice feature since it allows us to add the required metadata in
the background without having to get users to upgrade projects.
Also because these are default values the `.csproj` files will NOT
be changed when a user alters the project for some other reason.
i.e they WONT end up will ALL the `<AndroidResource />` items
having this new data on them.

The other thing to note is that these changes will only apply to the
Design-Time Builds. The target we are calling in this case is
`UpdateGeneratedFiles`. This target is not part of the `Build` chain.
As a result `_SetupDesingTimeBuildForCompile` will be called. This
sets `$(DesignTimeBuild)` to True. Calling this target will ONLY
update the DTB related files.

[0]: https://mhut.ch/journal/2015/06/30/build_time_code_generation_msbuild
[1]: https://docs.microsoft.com/en-us/visualstudio/msbuild/itemdefinitiongroup-element-msbuild