Don't throw exception on HTTP status 300 and no Location header
Fixes https://github.com/xamarin/xamarin-android/issues/1334

As per https://tools.ietf.org/html/rfc7231#section-6.4.1 the reponse isn't
required to contain the Location header and the client should act accordingly.
Since it is not documented what the action in this case should be, we're
following what Xamarin.iOS does and simply return the content of the request
as if it wasn't a redirect.