[build] Simplify building an API level other than default (#1353)
This commits adds the `$(API_LEVEL)` make variable which, when set,
will build Xamarin.Android and `Mono.Android.dll` for the specified
API level. It takes advantage of the variables found in
`build-tools/scripts/BuildEverything.mk` by indexing the
`$(ALL_FRAMEWORKS)` variable with the API level number and setting
the MSBuild properties accordingly.